requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
autopep8==2.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==1.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cmake==3.28.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dj-database-url==2.0.0
Django==3.2.19
django-cors-headers==4.3.1
django-filter==23.4
django-seed==0.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dm-tree==0.1.8
docx2txt==0.8
easyocr==1.7.0
et-xmlfile==1.1.0
executing==2.0.1
fake-factory==9999.9.9
filelock==3.13.1
filterpy==1.4.5
flatbuffers==24.3.7
fonttools==4.45.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
html-table-parser-python3==0.3.1
html5lib==0.9999999
httpcore==1.0.4
httpx==0.27.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.1.0
keyboard==0.13.5
kiwisolver==1.4.5
labelImg==1.8.6
lazy_loader==0.3
libclang==18.1.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.7
ndg-httpsclient==0.5.1
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.13.3
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.1.4
parso==0.8.3
Pillow==9.5.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.7
psycopg2==2.9.9
pure-eval==0.2.2
pusher==3.3.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-bidi==0.4.2
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-docx==1.1.0
pytweening==1.2.0
pytz==2023.3.post1
pyvi==0.1.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
setuptools==69.2.0
shapely==2.0.3
six==1.16.0
sklearn-crfsuite==0.3.6
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2023.3
ultralytics==8.0.114
unicodeit==0.7.5
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
youtube-transcript-api==0.6.2
